refactor: moving keepers to a new struct (#697)
