+++ app/keepers/keys.go
@@ -1,0 +1,48 @@
+package keepers
+
+import (
+	storetypes "cosmossdk.io/store/types"
+	paramstypes "github.com/cosmos/cosmos-sdk/x/params/types"
+)
+
+// GetSubspace gets existing substore from keeper.
+func (appKeepers *AppKeepers) GetSubspace(moduleName string) paramstypes.Subspace {
+	subspace, _ := appKeepers.ParamsKeeper.GetSubspace(moduleName)
+	return subspace
+}
+
+// GetKVStoreKey gets KV Store keys.
+func (appKeepers *AppKeepers) GetKVStoreKeys() map[string]*storetypes.KVStoreKey {
+	return appKeepers.keys
+}
+
+// GetTransientStoreKey gets Transient Store keys.
+func (appKeepers *AppKeepers) GetTransientStoreKeys() map[string]*storetypes.TransientStoreKey {
+	return appKeepers.tkeys
+}
+
+// GetMemoryStoreKey get memory Store keys.
+func (appKeepers *AppKeepers) GetMemoryStoreKeys() map[string]*storetypes.MemoryStoreKey {
+	return appKeepers.memKeys
+}
+
+// GetKey returns the KVStoreKey for the provided store key.
+//
+// NOTE: This is solely to be used for testing purposes.
+func (appKeepers *AppKeepers) GetKey(storeKey string) *storetypes.KVStoreKey {
+	return appKeepers.keys[storeKey]
+}
+
+// GetTKey returns the TransientStoreKey for the provided store key.
+//
+// NOTE: This is solely to be used for testing purposes.
+func (appKeepers *AppKeepers) GetTKey(storeKey string) *storetypes.TransientStoreKey {
+	return appKeepers.tkeys[storeKey]
+}
+
+// GetMemKey returns the MemStoreKey for the provided mem key.
+//
+// NOTE: This is solely used for testing purposes.
+func (appKeepers *AppKeepers) GetMemKey(storeKey string) *storetypes.MemoryStoreKey {
+	return appKeepers.memKeys[storeKey]
+}
