+++ app/keepers/utils.go
@@ -1,4 +1,4 @@
-package app
+package keepers
 
 import (
 	"bytes"
