+++ cmd/babylond/cmd/root.go
@@ -8,6 +8,7 @@ import (
 	confixcmd "cosmossdk.io/tools/confix/cmd"
 	"github.com/CosmWasm/wasmd/x/wasm"
 	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
+	appkeepers "github.com/babylonchain/babylon/app/keepers"
 	cmtcfg "github.com/cometbft/cometbft/config"
 	cmtcli "github.com/cometbft/cometbft/libs/cli"
 	dbm "github.com/cosmos/cosmos-db"
@@ -274,7 +275,7 @@ func newApp(logger log.Logger, db dbm.DB, traceStore io.Writer, appOpts serverty
 	}
 
 	homeDir := cast.ToString(appOpts.Get(flags.FlagHome))
-	privSigner, err := app.InitPrivSigner(homeDir)
+	privSigner, err := appkeepers.InitPrivSigner(homeDir)
 	if err != nil {
 		panic(err)
 	}
@@ -313,7 +314,7 @@ func appExport(
 		return servertypes.ExportedApp{}, errors.New("application home not set")
 	}
 
-	privSigner, err := app.InitPrivSigner(homePath)
+	privSigner, err := appkeepers.InitPrivSigner(homePath)
 	if err != nil {
 		panic(err)
 	}
