+++ cmd/babylond/cmd/testnet.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"cosmossdk.io/math"
+	appkeepers "github.com/babylonchain/babylon/app/keepers"
 	cmtconfig "github.com/cometbft/cometbft/config"
 	cmtos "github.com/cometbft/cometbft/libs/os"
 	cmttime "github.com/cometbft/cometbft/types/time"
@@ -35,7 +36,6 @@ import (
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/spf13/cobra"
 
-	"github.com/babylonchain/babylon/app"
 	appparams "github.com/babylonchain/babylon/app/params"
 	"github.com/babylonchain/babylon/privval"
 	"github.com/babylonchain/babylon/testutil/datagen"
@@ -321,7 +321,7 @@ func InitTestnet(
 		srvconfig.WriteConfigFile(filepath.Join(nodeDir, "config/app.toml"), babylonConfig)
 
 		// create and save client config
-		if _, err = app.CreateClientConfig(chainID, keyringBackend, nodeDir); err != nil {
+		if _, err = appkeepers.CreateClientConfig(chainID, keyringBackend, nodeDir); err != nil {
 			return err
 		}
 	}
