+++ test/e2e/initialization/init.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/babylonchain/babylon/app"
+	appkeepers "github.com/babylonchain/babylon/app/keepers"
 	"github.com/babylonchain/babylon/test/e2e/util"
 )
 
@@ -42,7 +42,7 @@ func InitChain(id, dataDir string, nodeConfigs []*NodeConfig, votingPeriod, expe
 	}
 
 	for _, node := range chain.nodes {
-		_, _ = app.CreateClientConfig(node.chain.chainMeta.Id, "test", node.configDir())
+		_, _ = appkeepers.CreateClientConfig(node.chain.chainMeta.Id, "test", node.configDir())
 	}
 
 	return chain.export(), nil
